Update README.md
This lead me to explore the backgrounds, experiences, and skills of different professional in Haiti.  For this project I want to study the profile of different professionals, and their evolution after graduation. I want to look at their background, the skills they have acquired, and the opportunities they have had.